Refactor reset password and email verify routes.
