+++ backend/app/routes/__init__.py
@@ -1,0 +1,2 @@
+from backend.app.routes.auth_routes import auth_bp
+from backend.app.routes.user_profile_routes import user_profile_bp
