+++ backend/app/routes/auth_routes.py
@@ -1,5 +1,5 @@
 from backend.app.models import User
-from backend.app.services import auth_service
+from backend.app.services import auth_service, email_confirm_service
 from flask import Blueprint, request, jsonify
 from flask_login import login_user, logout_user, login_required, current_user
 
@@ -25,15 +25,6 @@ def logout():
     logout_user()
     return jsonify({'message': 'Logged out successfully.'}), 200
 
-@auth_bp.route('/profile', methods=['GET'])
-@login_required
-def profile():
-    return jsonify({
-        'name': current_user.name,
-        'surname': current_user.surname,
-        'email': current_user.email
-    }), 200
-
 @auth_bp.route('/auth/google')
 def google_login():
     return auth_service.initiate_google_login()
@@ -45,4 +36,13 @@ def google_callback():
         login_user(user)
     return jsonify({'message': message}), status
 
+@auth_bp.route('/confirm_email/<token>', methods=['GET'])
+def confirm_email(token):
+    user = email_confirm_service.verify_email_token(token)
+    if not user:
+        return jsonify({'message': 'The token is invalid or expired.'}), 400
+
+    user.verify_email()
+
+    return jsonify({'message': 'Your email has been confirmed.'}), 200
