+++ backend/app/routes/user_profile_routes.py
@@ -2,9 +2,9 @@ from flask import Blueprint, request, jsonify
 from backend.app.models import User
 from backend.app.services.password_reset_service import send_password_reset_email
 
-password_reset_bp = Blueprint('password_reset', __name__)
+user_profile_bp = Blueprint('user_profile', __name__)
 
-@password_reset_bp.route('/reset_password', methods=['POST'])
+@user_profile_bp.route('/reset_password', methods=['POST'])
 def reset_password_request():
     data = request.get_json()
     user = User.query.filter_by(email=data['email']).first()
